refactor(docs): extract _write_auto_examples_root_index into a Sphinx extension
Moves _write_auto_examples_root_index (event: builder-inited, priority 600)
from conf.py into docs/source/_extensions/auto_examples_index.py. Builds
the top-level generated/auto_examples/index.rst aggregator by reading
sphinx-gallery's per-leaf index.rst files and stitching their thumbnail
grids into one parent page with hero intro and per-section card blocks.
Helpers moved alongside: _extract_sg_thumbnail_block and _section_block
(lifted to module scope to satisfy no-nested-functions). conf.py shrinks
from 4423 to 4096 LOC.

Part of C1 in docs_pipeline_architecture_review.md
